+++ packages/plugins/Filter/src/store/index.ts
@@ -3,6 +3,7 @@ import {
   generateSimpleMutations,
 } from '@repositoryname/vuex-generators'
 import { FilterConfiguration, PolarModule } from '@polar/lib-custom-types'
+import ClusterSource from 'ol/source/Cluster'
 import ChooseTimeFrame from '../components/ChooseTimeFrame.vue'
 import {
   FilterGetters,
@@ -11,7 +12,10 @@ import {
   LayerId,
   TargetProperty,
 } from '../types'
-import { updateFeatureVisibility } from '../utils/updateFeatureVisibility'
+import {
+  getLayer,
+  updateFeatureVisibility,
+} from '../utils/updateFeatureVisibility'
 import { setState } from '../utils/setState'
 import { arrayOnlyContains } from '../utils/arrayOnlyContains'
 import { parseTimeOption } from '../utils/parseTimeOption'
@@ -28,7 +32,17 @@ export const makeStoreModule = () => {
     namespaced: true,
     state: getInitialState(),
     actions: {
-      setupModule({ getters: { filterConfiguration }, commit }): void {
+      setupModule({
+        getters: { filterConfiguration },
+        rootGetters: { map },
+        commit,
+        dispatch,
+      }): void {
+        if (Object.entries(filterConfiguration.layers).length === 0) {
+          console.error(
+            '@polar/plugin-filter: No configuration for parameter "layers" was found. Plugin will not be usable.'
+          )
+        }
         Object.entries(filterConfiguration.layers).forEach(
           ([layerId, { categories, time }]) => {
             if (categories) {
@@ -54,6 +68,17 @@ export const makeStoreModule = () => {
                 },
               })
             }
+            // apply filter effects on layer loads
+            // @ts-expect-error | only layers with getSource allowed
+            let source = getLayer(map, layerId).getSource()
+            while (source instanceof ClusterSource) {
+              source = source.getSource()
+            }
+            source.on('featuresloadend', () =>
+              dispatch('updateFeatureVisibility', layerId)
+            )
+            // initially update visibility in case loading already took place
+            dispatch('updateFeatureVisibility', layerId)
           }
         )
       },