+++ packages/plugins/Export/CHANGELOG.md
@@ -1,8 +1,10 @@
 # CHANGELOG
 
-## unreleased
+## unpublished
 
 - Fix: PDF export with download:false now works correctly.
+- Fix: Adjust documentation to properly describe optionality of configuration parameters.
+- Chore: Update dependencies to latest versions.
 
 ## 1.2.0