+++ packages/plugins/Zoom/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## unpublished
+
+- Refactor: Remove redundant prop `isHorizontal` from `Zoom` component. Logic is now placed in `@polar/core`.
+
 ## 1.2.0
 
 - Feature: Improved implementation to make plugin SPA-ready.